q4_document_chunking/requirements.txt
# Core dependencies
langchain
langchain-community
langchain-openai
langchain-text-splitters

# Document processing
pypdf==3.17.4
python-docx==1.1.0
markdown==3.5.2
beautifulsoup4==4.12.2
lxml==4.9.3

# ML and NLP
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3

# Vector stores
chromadb==0.4.22
faiss-cpu==1.7.4

# Web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
tiktoken==0.5.2
tqdm==4.66.1
matplotlib==3.7.2
seaborn==0.12.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0 